@@ -35,8 +35,8 @@
         "wepack-cli": "^0.0.1-security"
     },
     "dependencies": {
-        "react": "^18.0.0",
-        "react-dom": "^18.0.0",
+        "react": "^18.1.0",
+        "react-dom": "^18.1.0",
         "react-icons": "^4.3.1",
         "styled-components": "^5.3.5"
     },
