@@ -1,5 +1,10 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+// import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import { App } from './App'
 
-ReactDOM.render(<App />, document.getElementById('app'))
+const container = document.getElementById('app')
+
+const root = createRoot(container)
+// ReactDOM.render(<App />, document.getElementById('app'))
+root.render(<App />)
