@@ -1,20 +1,41 @@
-import React from 'react'
+import React, { Fragment, useEffect, useRef, useState } from 'react'
 import { MdFavoriteBorder } from 'react-icons/md'
-import { Button, Img, ImgWrapper } from './styles'
+import { Article, Button, Img, ImgWrapper } from './styles'
 
-const DEFAULT_IMAGE = 'https://res.cloudinary.com/midudev/image/upload/w_300/q_80/v1560262103/dogs.png'
+const DEFAULT_IMAGE = 'https://images.unsplash.com/photo-1518791841217-8f162f1e1131?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=60'
 
 export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
+  const element = useRef(null)
+  const [show, setShow] = useState(false)
+
+  useEffect(function () {
+    const observer = new window.IntersectionObserver(function (entries) {
+      const { isIntersecting } = entries[0]
+      console.log(isIntersecting)
+      if (isIntersecting) {
+        setShow(true)
+        observer.disconnect()
+      }
+    })
+    observer.observe(element.current)
+  }, [element])
+
   return (
-    <article>
-      <a href={`/detail/${id}`}>
-        <ImgWrapper>
-          <Img src={src} />
-        </ImgWrapper>
-      </a>
-      <Button>
-        <MdFavoriteBorder size='32px' /> {likes} likes !
-      </Button>
-    </article>
+    <Article ref={element}>
+      {
+        show &&
+          <>
+            <a href={`/detail/${id}`}>
+              <ImgWrapper>
+                <Img src={src} />
+              </ImgWrapper>
+            </a>
+
+            <Button>
+              <MdFavoriteBorder size='32px' /> {likes} likes!
+            </Button>
+          </>
+      }
+    </Article>
   )
 }
